Create Updated HTML
Added a pulsing animation to my clock made everything a little more appealing and easier to read. Added a what I like about earth section that scrolls, and a place  holder for submissions.
I will need to start refactoring my code, it is getting too long and need to cut it into different files, but ill address that at a later date.