Create Updated CSS style
CSS was where a lot more stuff happened, also made font easier to read, and accounted for different web browsers when it comes to animations for the pulsation I added. The table, while still not a lot of information, is centered and bigger.  All in all this is just a better improvement on my previous code visually, especially the text size.