Create Earth day webpage 3.0
added some horizontal navigation and the CSS structure to go with it, along with links to home, an about earth day, and contact, they don't lead to any real sites, and I'm debating on making one just to include some of the remaining features, as this is now getting to 200 lines, I'm also reading about restructuring and how to approach implementation, so that I can do that from the start next time, It seems CSS you want to just structure very well, along with HTML, and call Java files to save who ever is reading the codes scrolling finger. In my final project, I will approach it in this manner.